Cleanup
* Fixed some typos
* Removed some no longer correct parameter documentation
* Made FileThreadService an interface and renamed the implementation to FileThreadServiceImpl.
* Changed to using @autowired instead of static references to the instance.
