+++ MovieFinder/src/main/java/edu/chalmers/dat076/moviefinder/listener/FileSystemListener.java
@@ -25,14 +25,12 @@ public interface FileSystemListener {
     /**
      * newFile is called when a new file is added to a directory that is being watched.
      * @param path The absolute path to the file. Includes the name of the file.
-     * @param name The name of the file.
      */
     public void newFile(Path path);
 
     /**
      * oldPath is called when a file is removed/renamed in a directory that is being watched.
      * @param path The absolute path to the file. Includes the name of the file.
-     * @param name The name of the file.
      */
     public void oldPath(Path path);
