+++ MovieFinder/src/main/java/edu/chalmers/dat076/moviefinder/service/MovieFileDatabaseHandler.java
@@ -14,11 +14,10 @@ public interface MovieFileDatabaseHandler {
     void updateFiles(List<Path> paths, Path basePath);
     
     /**
-     * Takes a file path and a file name and attempts to store it in the database.
+     * Takes a file path and attempts to store it in the database.
      * This operation can be unsuccessful for various reasons, if the file is decided to already exist
      * in the database, or if the file name can not be parsed.
      * @param path The absolute file path of the file. Including the file name.
-     * @param name The file name.
      */
     void saveFile(Path path);
 
@@ -26,7 +25,6 @@ public interface MovieFileDatabaseHandler {
      * Takes a file path and a file name of a removed file and attempts to remove any references to this file
      * from the database.
      * @param path The absolute file path of the file. Including the file name.
-     * @return True if the file was removed, else false.
      */
     void removeFile(Path path);
 }
